+++ Tests.GoogleDrive/StorageTests.cs
@@ -15,7 +15,7 @@ namespace Tests.GoogleDrive
         [TestMethod]
         public async Task CreateFolder_IsSuccess()
         {
-            var action = new StorageActions(InvocationContext, FileManager);
+            var action = new FolderActions(InvocationContext, FileManager);
             var input = new CreateFolderRequest { FolderName = "TestFolder", ParentFolderId = "10Ugy3Y7-kSXNtxFhnESuVpg89WTGl8wB" };
             var result =  action.CreateFolder(input);
             Console.WriteLine(result.FolderID);
@@ -26,7 +26,7 @@ namespace Tests.GoogleDrive
         [TestMethod]
         public async Task CheckFolderExists_IsSuccess()
         {
-            var action = new StorageActions(InvocationContext, FileManager);
+            var action = new FolderActions(InvocationContext, FileManager);
             var input = new CheckFolderRequest { FolderName = "TestFolder", ParentFolderId = "10Ugy3Y7-kSXNtxFhnESuVpg89WTGl8wB" };
             var result = await action.CheckFolderExists(input);
